Refact tests
