+++ NetTelebot.Tests/NetTelebot.Tests.csproj
@@ -73,13 +73,14 @@
   <ItemGroup>
     <Compile Include="MockServer\MockServer.cs" />
     <Compile Include="MockServer\ResponseString.cs" />
+    <Compile Include="NullReferenceExceptionTest\MessageInfoSimpleTypesTest.cs" />
     <Compile Include="RequestToTelegramTest\TelegramBotInlineKeyboardTest.cs" />
     <Compile Include="RequestToTelegramTest\TelegramBotKeyboardTest.cs" />
     <Compile Include="ResponseTest\BooleanResultParserTest.cs" />
     <Compile Include="ResponseTest\CallbackQueryParserTest.cs" />
     <Compile Include="ResponseTest\GetUpdatesResultParserTest.cs" />
     <Compile Include="ResponseTest\MeInfoParserTest.cs" />
-    <Compile Include="MessageInfoNullRefExcTest.cs" />
+    <Compile Include="NullReferenceExceptionTest\MessageInfoTest.cs" />
     <Compile Include="ResponseTest\MessageInfoParserTest.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="RequestToMockTest\TelegramBotInlineKeyboardTest.cs" />
